examples/openai-agents/requirements.txt
openai>=1.99.1
openai-agents>=0.2.4
python-dotenv>=1.0.0
requests>=2.32.0
pydantic>=2.7.0
httpx>=0.27.0